requirements.txt
PyQt5
pymongo
python-dotenv
bcrypt